list saved R objects

list saved R objects

list_objects(
  objects_path = jamsession_paths()$objects,
  most_recent = FALSE,
  object_prefix = "",
  object_suffix = "[.]RData$",
  add_stats = FALSE,
  ...
)

Arguments

objects_path
character vector of one or more file paths to search for saved R objects. When objects_path=NULL, it uses the output from jamsession_paths()$objects.

most_recent
logical whether to return only the most recent saved entry for each session name. When most_recent is FALSE, this function returns all saved versions for all R session names.

object_prefix, object_suffix
character string used as prefix or suffix when searching each path in objects_path for matching file names. By default, files are expected to have file extension .RData.

add_stats
logical indicating whether to include summary stats for each object.

...
additional arguments are ignored.

Value

list containing

object_df

data.frame of R objects

object

vector of matching R object names

Details

This function lists saved R objects for the given objects_path, usually called by grep_objects().

When most_recent=TRUE it returns only the most recent version of each file.

It is possible to supply multiple file paths using the argument objects_path, in which case all objects are returned for all directories, then they are sorted by save_date. The input objects_path is stored as an ordered factor, such that two files with the same save_date will still be ordered consistent with the input directories in objects_path. In this case, when most_recent=TRUE there will be one entry returned per object name, and the first entry will be the most recent entry ordered by save_date then objects_path. In practice, it is recommended to use one objects_path location, but this option exists.

For example, load_object(grep_objects("my_favorite_object")) will load the most recent object by save_date where any ties prefer the order in objects_path. Similarly, load_object("my_favorite_object") does the same thing, without the grep_patten search step.

See also

Examples

# by default the full list includes all past versions
head(list_objects());

#> $object_df
#>                                                                object save_date
#> all_samples_sle2_07jan2021                           all_samples_sle2 07jan2021
#> arris_df_01jul2020                                           arris_df 01jul2020
#> arris_df_01jun2020                                           arris_df 01jun2020
#> arris_df_02jul2020                                           arris_df 02jul2020
#> arris_df_02jun2020                                           arris_df 02jun2020
#> arris_df_03jul2020                                           arris_df 03jul2020
#> arris_df_03jun2020                                           arris_df 03jun2020
#> arris_df_04jul2020                                           arris_df 04jul2020
#> arris_df_04jun2020                                           arris_df 04jun2020
#> arris_df_05jul2020                                           arris_df 05jul2020
#> arris_df_05jun2020                                           arris_df 05jun2020
#> arris_df_05may2020                                           arris_df 05may2020
#> arris_df_06jul2020                                           arris_df 06jul2020
#> arris_df_06jun2020                                           arris_df 06jun2020
#> arris_df_06may2020                                           arris_df 06may2020
#> arris_df_07jul2020                                           arris_df 07jul2020
#> arris_df_07jun2020                                           arris_df 07jun2020
#> arris_df_07may2020                                           arris_df 07may2020
#> arris_df_08jul2020                                           arris_df 08jul2020
#> arris_df_08jun2020                                           arris_df 08jun2020
#> arris_df_08may2020                                           arris_df 08may2020
#> arris_df_09jul2020                                           arris_df 09jul2020
#> arris_df_09jun2020                                           arris_df 09jun2020
#> arris_df_09may2020                                           arris_df 09may2020
#> arris_df_10jul2020                                           arris_df 10jul2020
#> arris_df_10jun2020                                           arris_df 10jun2020
#> arris_df_10may2020                                           arris_df 10may2020
#> arris_df_11jul2020                                           arris_df 11jul2020
#> arris_df_11jun2020                                           arris_df 11jun2020
#> arris_df_11may2020                                           arris_df 11may2020
#> arris_df_12jul2020                                           arris_df 12jul2020
#> arris_df_12jun2020                                           arris_df 12jun2020
#> arris_df_12may2020                                           arris_df 12may2020
#> arris_df_13jul2020                                           arris_df 13jul2020
#> arris_df_13jun2020                                           arris_df 13jun2020
#> arris_df_13may2020                                           arris_df 13may2020
#> arris_df_14jul2020                                           arris_df 14jul2020
#> arris_df_14jun2020                                           arris_df 14jun2020
#> arris_df_14may2020                                           arris_df 14may2020
#> arris_df_15jul2020                                           arris_df 15jul2020
#> arris_df_15jun2020                                           arris_df 15jun2020
#> arris_df_15may2020                                           arris_df 15may2020
#> arris_df_16jul2020                                           arris_df 16jul2020
#> arris_df_16jun2020                                           arris_df 16jun2020
#> arris_df_16may2020                                           arris_df 16may2020
#> arris_df_17jul2020                                           arris_df 17jul2020
#> arris_df_17jun2020                                           arris_df 17jun2020
#> arris_df_17may2020                                           arris_df 17may2020
#> arris_df_18jul2020                                           arris_df 18jul2020
#> arris_df_18jun2020                                           arris_df 18jun2020
#> arris_df_18may2020                                           arris_df 18may2020
#> arris_df_19jul2020                                           arris_df 19jul2020
#> arris_df_19jun2020                                           arris_df 19jun2020
#> arris_df_19may2020                                           arris_df 19may2020
#> arris_df_20jul2020                                           arris_df 20jul2020
#> arris_df_20jun2020                                           arris_df 20jun2020
#> arris_df_20may2020                                           arris_df 20may2020
#> arris_df_21jul2020                                           arris_df 21jul2020
#> arris_df_21jun2020                                           arris_df 21jun2020
#> arris_df_21may2020                                           arris_df 21may2020
#> arris_df_22jul2020                                           arris_df 22jul2020
#> arris_df_22jun2020                                           arris_df 22jun2020
#> arris_df_22may2020                                           arris_df 22may2020
#> arris_df_23jul2020                                           arris_df 23jul2020
#> arris_df_23jun2020                                           arris_df 23jun2020
#> arris_df_23may2020                                           arris_df 23may2020
#> arris_df_24jul2020                                           arris_df 24jul2020
#> arris_df_24jun2020                                           arris_df 24jun2020
#> arris_df_24may2020                                           arris_df 24may2020
#> arris_df_25jul2020                                           arris_df 25jul2020
#> arris_df_25jun2020                                           arris_df 25jun2020
#> arris_df_25may2020                                           arris_df 25may2020
#> arris_df_26jul2020                                           arris_df 26jul2020
#> arris_df_26jun2020                                           arris_df 26jun2020
#> arris_df_26may2020                                           arris_df 26may2020
#> arris_df_27jul2020                                           arris_df 27jul2020
#> arris_df_27jun2020                                           arris_df 27jun2020
#> arris_df_27may2020                                           arris_df 27may2020
#> arris_df_28jun2020                                           arris_df 28jun2020
#> arris_df_28may2020                                           arris_df 28may2020
#> arris_df_29jun2020                                           arris_df 29jun2020
#> arris_df_29may2020                                           arris_df 29may2020
#> arris_df_30jun2020                                           arris_df 30jun2020
#> arris_df_30may2020                                           arris_df 30may2020
#> arris_df_31may2020                                           arris_df 31may2020
#> arris_df-irow_11jun2020                                 arris_df-irow 11jun2020
#> count_test_24nov2020                                       count_test 24nov2020
#> dietrich_track_lines_l_10apr2020               dietrich_track_lines_l 10apr2020
#> dietrichTxiNames_20mar2020                           dietrichTxiNames 20mar2020
#> dietrichTxis_20mar2020                                   dietrichTxis 20mar2020
#> dm_ab_list_11aug2020                                       dm_ab_list 11aug2020
#> dm_cbc_list_11aug2020                                     dm_cbc_list 11aug2020
#> dm_jdm_samples_df_12aug2020                         dm_jdm_samples_df 12aug2020
#> dmjdm_ipa_pathways_30jul2020                       dmjdm_ipa_pathways 30jul2020
#> gokey_data_06jul2020                                       gokey_data 06jul2020
#> gokeynmatlist250k_11feb2021                         gokeynmatlist250k 11feb2021
#> jam_divergent_17nov2020                                 jam_divergent 17nov2020
#> jam_linear_17nov2020                                       jam_linear 17nov2020
#> mem_dmlupus_affysoma_plots_12apr2020       mem_dmlupus_affysoma_plots 12apr2020
#> my_df_12jun2020                                                 my_df 12jun2020
#> my_df-my_df2_12jun2020                                   my_df-my_df2 12jun2020
#> proteinatlas_expr_fdb11_23jul2020             proteinatlas_expr_fdb11 23jul2020
#> proteinatlas_expression_fdb11_23jul2020 proteinatlas_expression_fdb11 23jul2020
#> proteinatlas_genesets_fdb11_23jul2020     proteinatlas_genesets_fdb11 23jul2020
#>                                         days_old file_size
#> all_samples_sle2_07jan2021                    50 314 bytes
#> arris_df_01jul2020                           240    546 kb
#> arris_df_01jun2020                           270    184 kb
#> arris_df_02jul2020                           239    560 kb
#> arris_df_02jun2020                           269    190 kb
#> arris_df_03jul2020                           238    588 kb
#> arris_df_03jun2020                           268    200 kb
#> arris_df_04jul2020                           237    596 kb
#> arris_df_04jun2020                           267    206 kb
#> arris_df_05jul2020                           236    616 kb
#> arris_df_05jun2020                           266    213 kb
#> arris_df_05may2020                           297    6.3 kb
#> arris_df_06jul2020                           235    631 kb
#> arris_df_06jun2020                           265    236 kb
#> arris_df_06may2020                           296     18 kb
#> arris_df_07jul2020                           234    637 kb
#> arris_df_07jun2020                           264    257 kb
#> arris_df_07may2020                           295     24 kb
#> arris_df_08jul2020                           233    645 kb
#> arris_df_08jun2020                           263    276 kb
#> arris_df_08may2020                           294     31 kb
#> arris_df_09jul2020                           232    659 kb
#> arris_df_09jun2020                           262    282 kb
#> arris_df_09may2020                           293     41 kb
#> arris_df_10jul2020                           231    664 kb
#> arris_df_10jun2020                           261    294 kb
#> arris_df_10may2020                           292     48 kb
#> arris_df_11jul2020                           230    685 kb
#> arris_df_11jun2020                           260    298 kb
#> arris_df_11may2020                           291     60 kb
#> arris_df_12jul2020                           229    699 kb
#> arris_df_12jun2020                           259    301 kb
#> arris_df_12may2020                           290     64 kb
#> arris_df_13jul2020                           228    716 kb
#> arris_df_13jun2020                           258    310 kb
#> arris_df_13may2020                           289     69 kb
#> arris_df_14jul2020                           227    720 kb
#> arris_df_14jun2020                           257    321 kb
#> arris_df_14may2020                           288     76 kb
#> arris_df_15jul2020                           226    733 kb
#> arris_df_15jun2020                           256    326 kb
#> arris_df_15may2020                           287     83 kb
#> arris_df_16jul2020                           225    743 kb
#> arris_df_16jun2020                           255    342 kb
#> arris_df_16may2020                           286     92 kb
#> arris_df_17jul2020                           224    746 kb
#> arris_df_17jun2020                           254    354 kb
#> arris_df_17may2020                           285    100 kb
#> arris_df_18jul2020                           223    763 kb
#> arris_df_18jun2020                           253    359 kb
#> arris_df_18may2020                           284    103 kb
#> arris_df_19jul2020                           222    769 kb
#> arris_df_19jun2020                           252    378 kb
#> arris_df_19may2020                           283    109 kb
#> arris_df_20jul2020                           221    775 kb
#> arris_df_20jun2020                           251    402 kb
#> arris_df_20may2020                           282    113 kb
#> arris_df_21jul2020                           220    790 kb
#> arris_df_21jun2020                           250    411 kb
#> arris_df_21may2020                           281    117 kb
#> arris_df_22jul2020                           219    799 kb
#> arris_df_22jun2020                           249    431 kb
#> arris_df_22may2020                           280    122 kb
#> arris_df_23jul2020                           218    805 kb
#> arris_df_23jun2020                           248    445 kb
#> arris_df_23may2020                           279    127 kb
#> arris_df_24jul2020                           217    825 kb
#> arris_df_24jun2020                           247    457 kb
#> arris_df_24may2020                           278    135 kb
#> arris_df_25jul2020                           216    842 kb
#> arris_df_25jun2020                           246    474 kb
#> arris_df_25may2020                           277    143 kb
#> arris_df_26jul2020                           215    848 kb
#> arris_df_26jun2020                           245    477 kb
#> arris_df_26may2020                           276    151 kb
#> arris_df_27jul2020                           214    852 kb
#> arris_df_27jun2020                           244    485 kb
#> arris_df_27may2020                           275    159 kb
#> arris_df_28jun2020                           243    501 kb
#> arris_df_28may2020                           274    161 kb
#> arris_df_29jun2020                           242    519 kb
#> arris_df_29may2020                           273    162 kb
#> arris_df_30jun2020                           241    535 kb
#> arris_df_30may2020                           272    170 kb
#> arris_df_31may2020                           271    183 kb
#> arris_df-irow_11jun2020                      260    294 kb
#> count_test_24nov2020                          94    539 kb
#> dietrich_track_lines_l_10apr2020             322      1 kb
#> dietrichTxiNames_20mar2020                   343 345 bytes
#> dietrichTxis_20mar2020                       343     51 Mb
#> dm_ab_list_11aug2020                         199    764 kb
#> dm_cbc_list_11aug2020                        199     22 kb
#> dm_jdm_samples_df_12aug2020                  198 475 bytes
#> dmjdm_ipa_pathways_30jul2020                 211     75 kb
#> gokey_data_06jul2020                         235    1.6 Mb
#> gokeynmatlist250k_11feb2021                   15    979 Mb
#> jam_divergent_17nov2020                      101    1.3 kb
#> jam_linear_17nov2020                         101 619 bytes
#> mem_dmlupus_affysoma_plots_12apr2020         320    299 kb
#> my_df_12jun2020                              259 160 bytes
#> my_df-my_df2_12jun2020                       259 212 bytes
#> proteinatlas_expr_fdb11_23jul2020            218      5 Mb
#> proteinatlas_expression_fdb11_23jul2020      218    5.3 Mb
#> proteinatlas_genesets_fdb11_23jul2020        218    109 kb
#>                                                              object_path
#> all_samples_sle2_07jan2021              /Users/wardjm/Projects/R-objects
#> arris_df_01jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_01jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_02jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_02jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_03jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_03jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_04jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_04jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_05jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_05jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_05may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_06jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_06jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_06may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_07jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_07jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_07may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_08jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_08jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_08may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_09jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_09jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_09may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_10jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_10jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_10may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_11jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_11jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_11may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_12jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_12jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_12may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_13jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_13jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_13may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_14jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_14jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_14may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_15jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_15jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_15may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_16jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_16jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_16may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_17jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_17jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_17may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_18jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_18jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_18may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_19jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_19jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_19may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_20jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_20jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_20may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_21jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_21jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_21may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_22jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_22jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_22may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_23jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_23jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_23may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_24jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_24jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_24may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_25jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_25jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_25may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_26jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_26jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_26may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_27jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df_27jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_27may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_28jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_28may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_29jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_29may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_30jun2020                      /Users/wardjm/Projects/R-objects
#> arris_df_30may2020                      /Users/wardjm/Projects/R-objects
#> arris_df_31may2020                      /Users/wardjm/Projects/R-objects
#> arris_df-irow_11jun2020                 /Users/wardjm/Projects/R-objects
#> count_test_24nov2020                    /Users/wardjm/Projects/R-objects
#> dietrich_track_lines_l_10apr2020        /Users/wardjm/Projects/R-objects
#> dietrichTxiNames_20mar2020              /Users/wardjm/Projects/R-objects
#> dietrichTxis_20mar2020                  /Users/wardjm/Projects/R-objects
#> dm_ab_list_11aug2020                    /Users/wardjm/Projects/R-objects
#> dm_cbc_list_11aug2020                   /Users/wardjm/Projects/R-objects
#> dm_jdm_samples_df_12aug2020             /Users/wardjm/Projects/R-objects
#> dmjdm_ipa_pathways_30jul2020            /Users/wardjm/Projects/R-objects
#> gokey_data_06jul2020                    /Users/wardjm/Projects/R-objects
#> gokeynmatlist250k_11feb2021             /Users/wardjm/Projects/R-objects
#> jam_divergent_17nov2020                 /Users/wardjm/Projects/R-objects
#> jam_linear_17nov2020                    /Users/wardjm/Projects/R-objects
#> mem_dmlupus_affysoma_plots_12apr2020    /Users/wardjm/Projects/R-objects
#> my_df_12jun2020                         /Users/wardjm/Projects/R-objects
#> my_df-my_df2_12jun2020                  /Users/wardjm/Projects/R-objects
#> proteinatlas_expr_fdb11_23jul2020       /Users/wardjm/Projects/R-objects
#> proteinatlas_expression_fdb11_23jul2020 /Users/wardjm/Projects/R-objects
#> proteinatlas_genesets_fdb11_23jul2020   /Users/wardjm/Projects/R-objects
#>                                                                           object_file
#> all_samples_sle2_07jan2021                           all_samples_sle2_07jan2021.RData
#> arris_df_01jul2020                                           arris_df_01jul2020.RData
#> arris_df_01jun2020                                           arris_df_01jun2020.RData
#> arris_df_02jul2020                                           arris_df_02jul2020.RData
#> arris_df_02jun2020                                           arris_df_02jun2020.RData
#> arris_df_03jul2020                                           arris_df_03jul2020.RData
#> arris_df_03jun2020                                           arris_df_03jun2020.RData
#> arris_df_04jul2020                                           arris_df_04jul2020.RData
#> arris_df_04jun2020                                           arris_df_04jun2020.RData
#> arris_df_05jul2020                                           arris_df_05jul2020.RData
#> arris_df_05jun2020                                           arris_df_05jun2020.RData
#> arris_df_05may2020                                           arris_df_05may2020.RData
#> arris_df_06jul2020                                           arris_df_06jul2020.RData
#> arris_df_06jun2020                                           arris_df_06jun2020.RData
#> arris_df_06may2020                                           arris_df_06may2020.RData
#> arris_df_07jul2020                                           arris_df_07jul2020.RData
#> arris_df_07jun2020                                           arris_df_07jun2020.RData
#> arris_df_07may2020                                           arris_df_07may2020.RData
#> arris_df_08jul2020                                           arris_df_08jul2020.RData
#> arris_df_08jun2020                                           arris_df_08jun2020.RData
#> arris_df_08may2020                                           arris_df_08may2020.RData
#> arris_df_09jul2020                                           arris_df_09jul2020.RData
#> arris_df_09jun2020                                           arris_df_09jun2020.RData
#> arris_df_09may2020                                           arris_df_09may2020.RData
#> arris_df_10jul2020                                           arris_df_10jul2020.RData
#> arris_df_10jun2020                                           arris_df_10jun2020.RData
#> arris_df_10may2020                                           arris_df_10may2020.RData
#> arris_df_11jul2020                                           arris_df_11jul2020.RData
#> arris_df_11jun2020                                           arris_df_11jun2020.RData
#> arris_df_11may2020                                           arris_df_11may2020.RData
#> arris_df_12jul2020                                           arris_df_12jul2020.RData
#> arris_df_12jun2020                                           arris_df_12jun2020.RData
#> arris_df_12may2020                                           arris_df_12may2020.RData
#> arris_df_13jul2020                                           arris_df_13jul2020.RData
#> arris_df_13jun2020                                           arris_df_13jun2020.RData
#> arris_df_13may2020                                           arris_df_13may2020.RData
#> arris_df_14jul2020                                           arris_df_14jul2020.RData
#> arris_df_14jun2020                                           arris_df_14jun2020.RData
#> arris_df_14may2020                                           arris_df_14may2020.RData
#> arris_df_15jul2020                                           arris_df_15jul2020.RData
#> arris_df_15jun2020                                           arris_df_15jun2020.RData
#> arris_df_15may2020                                           arris_df_15may2020.RData
#> arris_df_16jul2020                                           arris_df_16jul2020.RData
#> arris_df_16jun2020                                           arris_df_16jun2020.RData
#> arris_df_16may2020                                           arris_df_16may2020.RData
#> arris_df_17jul2020                                           arris_df_17jul2020.RData
#> arris_df_17jun2020                                           arris_df_17jun2020.RData
#> arris_df_17may2020                                           arris_df_17may2020.RData
#> arris_df_18jul2020                                           arris_df_18jul2020.RData
#> arris_df_18jun2020                                           arris_df_18jun2020.RData
#> arris_df_18may2020                                           arris_df_18may2020.RData
#> arris_df_19jul2020                                           arris_df_19jul2020.RData
#> arris_df_19jun2020                                           arris_df_19jun2020.RData
#> arris_df_19may2020                                           arris_df_19may2020.RData
#> arris_df_20jul2020                                           arris_df_20jul2020.RData
#> arris_df_20jun2020                                           arris_df_20jun2020.RData
#> arris_df_20may2020                                           arris_df_20may2020.RData
#> arris_df_21jul2020                                           arris_df_21jul2020.RData
#> arris_df_21jun2020                                           arris_df_21jun2020.RData
#> arris_df_21may2020                                           arris_df_21may2020.RData
#> arris_df_22jul2020                                           arris_df_22jul2020.RData
#> arris_df_22jun2020                                           arris_df_22jun2020.RData
#> arris_df_22may2020                                           arris_df_22may2020.RData
#> arris_df_23jul2020                                           arris_df_23jul2020.RData
#> arris_df_23jun2020                                           arris_df_23jun2020.RData
#> arris_df_23may2020                                           arris_df_23may2020.RData
#> arris_df_24jul2020                                           arris_df_24jul2020.RData
#> arris_df_24jun2020                                           arris_df_24jun2020.RData
#> arris_df_24may2020                                           arris_df_24may2020.RData
#> arris_df_25jul2020                                           arris_df_25jul2020.RData
#> arris_df_25jun2020                                           arris_df_25jun2020.RData
#> arris_df_25may2020                                           arris_df_25may2020.RData
#> arris_df_26jul2020                                           arris_df_26jul2020.RData
#> arris_df_26jun2020                                           arris_df_26jun2020.RData
#> arris_df_26may2020                                           arris_df_26may2020.RData
#> arris_df_27jul2020                                           arris_df_27jul2020.RData
#> arris_df_27jun2020                                           arris_df_27jun2020.RData
#> arris_df_27may2020                                           arris_df_27may2020.RData
#> arris_df_28jun2020                                           arris_df_28jun2020.RData
#> arris_df_28may2020                                           arris_df_28may2020.RData
#> arris_df_29jun2020                                           arris_df_29jun2020.RData
#> arris_df_29may2020                                           arris_df_29may2020.RData
#> arris_df_30jun2020                                           arris_df_30jun2020.RData
#> arris_df_30may2020                                           arris_df_30may2020.RData
#> arris_df_31may2020                                           arris_df_31may2020.RData
#> arris_df-irow_11jun2020                                 arris_df-irow_11jun2020.RData
#> count_test_24nov2020                                       count_test_24nov2020.RData
#> dietrich_track_lines_l_10apr2020               dietrich_track_lines_l_10apr2020.RData
#> dietrichTxiNames_20mar2020                           dietrichTxiNames_20mar2020.RData
#> dietrichTxis_20mar2020                                   dietrichTxis_20mar2020.RData
#> dm_ab_list_11aug2020                                       dm_ab_list_11aug2020.RData
#> dm_cbc_list_11aug2020                                     dm_cbc_list_11aug2020.RData
#> dm_jdm_samples_df_12aug2020                         dm_jdm_samples_df_12aug2020.RData
#> dmjdm_ipa_pathways_30jul2020                       dmjdm_ipa_pathways_30jul2020.RData
#> gokey_data_06jul2020                                       gokey_data_06jul2020.RData
#> gokeynmatlist250k_11feb2021                         gokeynmatlist250k_11feb2021.RData
#> jam_divergent_17nov2020                                 jam_divergent_17nov2020.RData
#> jam_linear_17nov2020                                       jam_linear_17nov2020.RData
#> mem_dmlupus_affysoma_plots_12apr2020       mem_dmlupus_affysoma_plots_12apr2020.RData
#> my_df_12jun2020                                                 my_df_12jun2020.RData
#> my_df-my_df2_12jun2020                                   my_df-my_df2_12jun2020.RData
#> proteinatlas_expr_fdb11_23jul2020             proteinatlas_expr_fdb11_23jul2020.RData
#> proteinatlas_expression_fdb11_23jul2020 proteinatlas_expression_fdb11_23jul2020.RData
#> proteinatlas_genesets_fdb11_23jul2020     proteinatlas_genesets_fdb11_23jul2020.RData
#> 
#> $object
#> NULL
#> 

# you can return only the most recent for each object
list_objects(most_recent=TRUE);

#> $object_df
#>                                                                object save_date
#> all_samples_sle2_07jan2021                           all_samples_sle2 07jan2021
#> arris_df_27jul2020                                           arris_df 27jul2020
#> arris_df-irow_11jun2020                                 arris_df-irow 11jun2020
#> count_test_24nov2020                                       count_test 24nov2020
#> dietrich_track_lines_l_10apr2020               dietrich_track_lines_l 10apr2020
#> dietrichTxiNames_20mar2020                           dietrichTxiNames 20mar2020
#> dietrichTxis_20mar2020                                   dietrichTxis 20mar2020
#> dm_ab_list_11aug2020                                       dm_ab_list 11aug2020
#> dm_cbc_list_11aug2020                                     dm_cbc_list 11aug2020
#> dm_jdm_samples_df_12aug2020                         dm_jdm_samples_df 12aug2020
#> dmjdm_ipa_pathways_30jul2020                       dmjdm_ipa_pathways 30jul2020
#> gokey_data_06jul2020                                       gokey_data 06jul2020
#> gokeynmatlist250k_11feb2021                         gokeynmatlist250k 11feb2021
#> jam_divergent_17nov2020                                 jam_divergent 17nov2020
#> jam_linear_17nov2020                                       jam_linear 17nov2020
#> mem_dmlupus_affysoma_plots_12apr2020       mem_dmlupus_affysoma_plots 12apr2020
#> my_df_12jun2020                                                 my_df 12jun2020
#> my_df-my_df2_12jun2020                                   my_df-my_df2 12jun2020
#> proteinatlas_expr_fdb11_23jul2020             proteinatlas_expr_fdb11 23jul2020
#> proteinatlas_expression_fdb11_23jul2020 proteinatlas_expression_fdb11 23jul2020
#> proteinatlas_genesets_fdb11_23jul2020     proteinatlas_genesets_fdb11 23jul2020
#>                                         days_old file_size
#> all_samples_sle2_07jan2021                    50 314 bytes
#> arris_df_27jul2020                           214    852 kb
#> arris_df-irow_11jun2020                      260    294 kb
#> count_test_24nov2020                          94    539 kb
#> dietrich_track_lines_l_10apr2020             322      1 kb
#> dietrichTxiNames_20mar2020                   343 345 bytes
#> dietrichTxis_20mar2020                       343     51 Mb
#> dm_ab_list_11aug2020                         199    764 kb
#> dm_cbc_list_11aug2020                        199     22 kb
#> dm_jdm_samples_df_12aug2020                  198 475 bytes
#> dmjdm_ipa_pathways_30jul2020                 211     75 kb
#> gokey_data_06jul2020                         235    1.6 Mb
#> gokeynmatlist250k_11feb2021                   15    979 Mb
#> jam_divergent_17nov2020                      101    1.3 kb
#> jam_linear_17nov2020                         101 619 bytes
#> mem_dmlupus_affysoma_plots_12apr2020         320    299 kb
#> my_df_12jun2020                              259 160 bytes
#> my_df-my_df2_12jun2020                       259 212 bytes
#> proteinatlas_expr_fdb11_23jul2020            218      5 Mb
#> proteinatlas_expression_fdb11_23jul2020      218    5.3 Mb
#> proteinatlas_genesets_fdb11_23jul2020        218    109 kb
#>                                                              object_path
#> all_samples_sle2_07jan2021              /Users/wardjm/Projects/R-objects
#> arris_df_27jul2020                      /Users/wardjm/Projects/R-objects
#> arris_df-irow_11jun2020                 /Users/wardjm/Projects/R-objects
#> count_test_24nov2020                    /Users/wardjm/Projects/R-objects
#> dietrich_track_lines_l_10apr2020        /Users/wardjm/Projects/R-objects
#> dietrichTxiNames_20mar2020              /Users/wardjm/Projects/R-objects
#> dietrichTxis_20mar2020                  /Users/wardjm/Projects/R-objects
#> dm_ab_list_11aug2020                    /Users/wardjm/Projects/R-objects
#> dm_cbc_list_11aug2020                   /Users/wardjm/Projects/R-objects
#> dm_jdm_samples_df_12aug2020             /Users/wardjm/Projects/R-objects
#> dmjdm_ipa_pathways_30jul2020            /Users/wardjm/Projects/R-objects
#> gokey_data_06jul2020                    /Users/wardjm/Projects/R-objects
#> gokeynmatlist250k_11feb2021             /Users/wardjm/Projects/R-objects
#> jam_divergent_17nov2020                 /Users/wardjm/Projects/R-objects
#> jam_linear_17nov2020                    /Users/wardjm/Projects/R-objects
#> mem_dmlupus_affysoma_plots_12apr2020    /Users/wardjm/Projects/R-objects
#> my_df_12jun2020                         /Users/wardjm/Projects/R-objects
#> my_df-my_df2_12jun2020                  /Users/wardjm/Projects/R-objects
#> proteinatlas_expr_fdb11_23jul2020       /Users/wardjm/Projects/R-objects
#> proteinatlas_expression_fdb11_23jul2020 /Users/wardjm/Projects/R-objects
#> proteinatlas_genesets_fdb11_23jul2020   /Users/wardjm/Projects/R-objects
#>                                                                           object_file
#> all_samples_sle2_07jan2021                           all_samples_sle2_07jan2021.RData
#> arris_df_27jul2020                                           arris_df_27jul2020.RData
#> arris_df-irow_11jun2020                                 arris_df-irow_11jun2020.RData
#> count_test_24nov2020                                       count_test_24nov2020.RData
#> dietrich_track_lines_l_10apr2020               dietrich_track_lines_l_10apr2020.RData
#> dietrichTxiNames_20mar2020                           dietrichTxiNames_20mar2020.RData
#> dietrichTxis_20mar2020                                   dietrichTxis_20mar2020.RData
#> dm_ab_list_11aug2020                                       dm_ab_list_11aug2020.RData
#> dm_cbc_list_11aug2020                                     dm_cbc_list_11aug2020.RData
#> dm_jdm_samples_df_12aug2020                         dm_jdm_samples_df_12aug2020.RData
#> dmjdm_ipa_pathways_30jul2020                       dmjdm_ipa_pathways_30jul2020.RData
#> gokey_data_06jul2020                                       gokey_data_06jul2020.RData
#> gokeynmatlist250k_11feb2021                         gokeynmatlist250k_11feb2021.RData
#> jam_divergent_17nov2020                                 jam_divergent_17nov2020.RData
#> jam_linear_17nov2020                                       jam_linear_17nov2020.RData
#> mem_dmlupus_affysoma_plots_12apr2020       mem_dmlupus_affysoma_plots_12apr2020.RData
#> my_df_12jun2020                                                 my_df_12jun2020.RData
#> my_df-my_df2_12jun2020                                   my_df-my_df2_12jun2020.RData
#> proteinatlas_expr_fdb11_23jul2020             proteinatlas_expr_fdb11_23jul2020.RData
#> proteinatlas_expression_fdb11_23jul2020 proteinatlas_expression_fdb11_23jul2020.RData
#> proteinatlas_genesets_fdb11_23jul2020     proteinatlas_genesets_fdb11_23jul2020.RData
#> 
#> $object
#> NULL
#> 

# add_stats makes sense when most_recent=TRUE
list_objects(most_recent=TRUE, add_stats=TRUE);

#> Warning: package ‘data.table’ was built under R version 3.6.2

#> $object_df
#>                           object save_date days_old file_size
#> 1               all_samples_sle2 07jan2021       50 314 bytes
#> 2                       arris_df 27jul2020      214    852 kb
#> 3                  arris_df-irow 11jun2020      260    294 kb
#> 4                     count_test 24nov2020       94    539 kb
#> 5         dietrich_track_lines_l 10apr2020      322      1 kb
#> 6               dietrichTxiNames 20mar2020      343 345 bytes
#> 7                   dietrichTxis 20mar2020      343     51 Mb
#> 8                     dm_ab_list 11aug2020      199    764 kb
#> 9                    dm_cbc_list 11aug2020      199     22 kb
#> 10             dm_jdm_samples_df 12aug2020      198 475 bytes
#> 11            dmjdm_ipa_pathways 30jul2020      211     75 kb
#> 12                    gokey_data 06jul2020      235    1.6 Mb
#> 13             gokeynmatlist250k 11feb2021       15    979 Mb
#> 14                 jam_divergent 17nov2020      101    1.3 kb
#> 15                    jam_linear 17nov2020      101 619 bytes
#> 16    mem_dmlupus_affysoma_plots 12apr2020      320    299 kb
#> 17                         my_df 12jun2020      259 160 bytes
#> 18                  my_df-my_df2 12jun2020      259 212 bytes
#> 19       proteinatlas_expr_fdb11 23jul2020      218      5 Mb
#> 20 proteinatlas_expression_fdb11 23jul2020      218    5.3 Mb
#> 21   proteinatlas_genesets_fdb11 23jul2020      218    109 kb
#>                         object_path
#> 1  /Users/wardjm/Projects/R-objects
#> 2  /Users/wardjm/Projects/R-objects
#> 3  /Users/wardjm/Projects/R-objects
#> 4  /Users/wardjm/Projects/R-objects
#> 5  /Users/wardjm/Projects/R-objects
#> 6  /Users/wardjm/Projects/R-objects
#> 7  /Users/wardjm/Projects/R-objects
#> 8  /Users/wardjm/Projects/R-objects
#> 9  /Users/wardjm/Projects/R-objects
#> 10 /Users/wardjm/Projects/R-objects
#> 11 /Users/wardjm/Projects/R-objects
#> 12 /Users/wardjm/Projects/R-objects
#> 13 /Users/wardjm/Projects/R-objects
#> 14 /Users/wardjm/Projects/R-objects
#> 15 /Users/wardjm/Projects/R-objects
#> 16 /Users/wardjm/Projects/R-objects
#> 17 /Users/wardjm/Projects/R-objects
#> 18 /Users/wardjm/Projects/R-objects
#> 19 /Users/wardjm/Projects/R-objects
#> 20 /Users/wardjm/Projects/R-objects
#> 21 /Users/wardjm/Projects/R-objects
#>                                      object_file number_saved total_size
#> 1               all_samples_sle2_07jan2021.RData            1  314 bytes
#> 2                       arris_df_27jul2020.RData           84      32 Mb
#> 3                  arris_df-irow_11jun2020.RData            1     294 kb
#> 4                     count_test_24nov2020.RData            1     539 kb
#> 5         dietrich_track_lines_l_10apr2020.RData            1       1 kb
#> 6               dietrichTxiNames_20mar2020.RData            1  345 bytes
#> 7                   dietrichTxis_20mar2020.RData            1      51 Mb
#> 8                     dm_ab_list_11aug2020.RData            1     764 kb
#> 9                    dm_cbc_list_11aug2020.RData            1      22 kb
#> 10             dm_jdm_samples_df_12aug2020.RData            1  475 bytes
#> 11            dmjdm_ipa_pathways_30jul2020.RData            1      75 kb
#> 12                    gokey_data_06jul2020.RData            1     1.6 Mb
#> 13             gokeynmatlist250k_11feb2021.RData            1     979 Mb
#> 14                 jam_divergent_17nov2020.RData            1     1.3 kb
#> 15                    jam_linear_17nov2020.RData            1  619 bytes
#> 16    mem_dmlupus_affysoma_plots_12apr2020.RData            1     299 kb
#> 17                         my_df_12jun2020.RData            1  160 bytes
#> 18                  my_df-my_df2_12jun2020.RData            1  212 bytes
#> 19       proteinatlas_expr_fdb11_23jul2020.RData            1       5 Mb
#> 20 proteinatlas_expression_fdb11_23jul2020.RData            1     5.3 Mb
#> 21   proteinatlas_genesets_fdb11_23jul2020.RData            1     109 kb
#>    days_span
#> 1          1
#> 2         84
#> 3          1
#> 4          1
#> 5          1
#> 6          1
#> 7          1
#> 8          1
#> 9          1
#> 10         1
#> 11         1
#> 12         1
#> 13         1
#> 14         1
#> 15         1
#> 16         1
#> 17         1
#> 18         1
#> 19         1
#> 20         1
#> 21         1
#> 
#> $object
#> NULL
#>