Search for saved R sessions

Search for saved R sessions

grep_jamsessions(
  pattern,
  sessions_path = jamsession_paths()$sessions,
  return_df = TRUE,
  sort_by_date = TRUE,
  ignore.case = TRUE,
  most_recent = TRUE,
  include_attrs = FALSE,
  add_stats = FALSE,
  verbose = FALSE,
  ...
)

Arguments

pattern
chracter pattern used with grep() to match session names.

sessions_path
character vector of one or more file paths to search for saved R sessions. When NULL, it uses the output from jamsession_paths()$sessions.

return_df
logical whether to return a vector (FALSE) or data.frame (TRUE) with detailed information about each session.

sort_by_date
logical whether to sort results by date, as opposed to the order retrieved from list.files.

ignore.case
logical sent to grep, by default TRUE which searches in case-insensitive mode.

most_recent
logical whether to return only the most recent saved version of each matching session, by default TRUE. This argument is passed to list_jamsessions().

include_attrs
logical indicating whether to include attributes "file_path", "file_size", "save_date", and "days_old", only used when return_df=FALSE.

add_stats
logical indicating whether to include summary stats for each session: number_saved is the number of past versions of the session; total_size is the total file size for all saved versions.

verbose
logical indicating whether to print verbose output.

...
additional arguments are passed to list_jamessions(), and jamba::provigrep(), as relevent. If session_prefix or session_suffix should be customized, it should be done here.

Details

This function searches for saved R sessions using one or more text or regular expression patterns.

See list_jamsessions() for more details about stored sessions.

This function can be chained with load_jamsession(), for example load_jamsession(grep_jamsessions("my_project")), and it will load the most recently saved R session.

See also

Examples

grep_jamsessions(".");

#>                         session save_date days_old file_size
#> LottyDah               LottyDah 12jun2020        0    1.1 kb
#> testOnly               testOnly 03jun2020        9     12 Mb
#> SARSinteractome SARSinteractome 27mar2020       77     12 Mb
#>                                      session_path
#> LottyDah        /Users/wardjm/Projects/R-sessions
#> testOnly        /Users/wardjm/Projects/R-sessions
#> SARSinteractome /Users/wardjm/Projects/R-sessions
#>                                               session_file
#> LottyDah               inProgress_LottyDah_12jun2020.RData
#> testOnly               inProgress_testOnly_03jun2020.RData
#> SARSinteractome inProgress_SARSinteractome_27mar2020.RData