List saved R sessions

List saved R sessions

list_jamsessions(
  sessions_path = jamsession_paths()$sessions,
  most_recent = FALSE,
  session_prefix = "^inProgress_",
  session_suffix = "[.]RData$",
  add_stats = FALSE,
  ...
)

Arguments

sessions_path
character vector of one or more file paths to search for saved R sessions. When NULL, it uses the output from jamsession_paths()$sessions.

most_recent
logical whether to return only the most recent saved entry for each session name. When most_recent is FALSE, this function returns all saved versions for all R session names.

session_prefix, session_suffix
character string used as prefix or suffix when searching each path in sessions_path for matching file names.

add_stats
logical indicating whether to include summary stats for each session: number_saved is the number of past versions of the session; total_size is the total file size for all saved versions.

...
additional parameters are ignored

Value

list containing

session_df

data.frame of R session information

session

vector of matching R session names

Details

This function lists available saved R sessions, usually called by grep_jamsessions().

When most_recent=TRUE it returns only the most recent session based upon the save_date.

It is possible to supply multiple file paths using the argument sessions_path, in which case all sessions are returned for all directories, then they are sorted by save_date. The input sessions_path is stored as an ordered factor, such that two sessions with the same save_date will still be ordered consistent with the input directories in sessions_path. In this case, when most_recent=TRUE there will be one entry returned per object name, and the first entry will be the most recent entry ordered by save_date then sessions_path. In practice, it is recommended to use one sessions_path location, but this option exists.

For example, jamsession(grep_sessions("my_session")) will load the most recent session by save_date where any ties prefer the order in sessions_path. Similarly, load_jamsession("my_session") does the same thing, without the grep_patten search step.

See also

Examples

list_jamsessions();

#> $session_df
#>                                           session save_date days_old file_size
#> adam19-rnaseq_09may2019             adam19-rnaseq 09may2019      659    231 Mb
#> bushel-mito-nuc-SNP_07aug2020 bushel-mito-nuc-SNP 07aug2020      203    1.1 Mb
#> bushel-mito-nuc-SNP_12jan2021 bushel-mito-nuc-SNP 12jan2021       45    2.3 Mb
#> bushel-mito-nuc-SNP_13jan2021 bushel-mito-nuc-SNP 13jan2021       44    2.3 Mb
#> bushel-mito-nuc-SNP_22sep2020 bushel-mito-nuc-SNP 22sep2020      157    2.3 Mb
#> dietrichmemfeb2021_23feb2021   dietrichmemfeb2021 23feb2021        3    151 Mb
#> DMdata_07jan2021                           DMdata 07jan2021       50    622 Mb
#> DMdata_07oct2020                           DMdata 07oct2020      142    590 Mb
#> DMdata_11feb2021                           DMdata 11feb2021       15    625 Mb
#> DMdata_16dec2020                           DMdata 16dec2020       72    618 Mb
#> gokeyH1PD29may2019_27jan2021   gokeyH1PD29may2019 27jan2021       30     11 Mb
#> gokeyH1PD29may2019_28jan2021   gokeyH1PD29may2019 28jan2021       29     12 Mb
#> LottyDah_12jun2020                       LottyDah 12jun2020      259    1.1 kb
#> proteinatlas_24jul2020               proteinatlas 24jul2020      217     59 Mb
#> ritux-workflow_10aug2020           ritux-workflow 10aug2020      200     35 Mb
#> SARSinteractome_24mar2020         SARSinteractome 24mar2020      339      7 Mb
#> SARSinteractome_26mar2020         SARSinteractome 26mar2020      337     12 Mb
#> SARSinteractome_27mar2020         SARSinteractome 27mar2020      336     12 Mb
#> slicejam-testing_07jul2020       slicejam-testing 07jul2020      234    5.8 Mb
#> testOnly_03jun2020                       testOnly 03jun2020      268     12 Mb
#>                                                    session_path
#> adam19-rnaseq_09may2019       /Users/wardjm/Projects/R-sessions
#> bushel-mito-nuc-SNP_07aug2020 /Users/wardjm/Projects/R-sessions
#> bushel-mito-nuc-SNP_12jan2021 /Users/wardjm/Projects/R-sessions
#> bushel-mito-nuc-SNP_13jan2021 /Users/wardjm/Projects/R-sessions
#> bushel-mito-nuc-SNP_22sep2020 /Users/wardjm/Projects/R-sessions
#> dietrichmemfeb2021_23feb2021  /Users/wardjm/Projects/R-sessions
#> DMdata_07jan2021              /Users/wardjm/Projects/R-sessions
#> DMdata_07oct2020              /Users/wardjm/Projects/R-sessions
#> DMdata_11feb2021              /Users/wardjm/Projects/R-sessions
#> DMdata_16dec2020              /Users/wardjm/Projects/R-sessions
#> gokeyH1PD29may2019_27jan2021  /Users/wardjm/Projects/R-sessions
#> gokeyH1PD29may2019_28jan2021  /Users/wardjm/Projects/R-sessions
#> LottyDah_12jun2020            /Users/wardjm/Projects/R-sessions
#> proteinatlas_24jul2020        /Users/wardjm/Projects/R-sessions
#> ritux-workflow_10aug2020      /Users/wardjm/Projects/R-sessions
#> SARSinteractome_24mar2020     /Users/wardjm/Projects/R-sessions
#> SARSinteractome_26mar2020     /Users/wardjm/Projects/R-sessions
#> SARSinteractome_27mar2020     /Users/wardjm/Projects/R-sessions
#> slicejam-testing_07jul2020    /Users/wardjm/Projects/R-sessions
#> testOnly_03jun2020            /Users/wardjm/Projects/R-sessions
#>                                                                 session_file
#> adam19-rnaseq_09may2019             inProgress_adam19-rnaseq_09may2019.RData
#> bushel-mito-nuc-SNP_07aug2020 inProgress_bushel-mito-nuc-SNP_07aug2020.RData
#> bushel-mito-nuc-SNP_12jan2021 inProgress_bushel-mito-nuc-SNP_12jan2021.RData
#> bushel-mito-nuc-SNP_13jan2021 inProgress_bushel-mito-nuc-SNP_13jan2021.RData
#> bushel-mito-nuc-SNP_22sep2020 inProgress_bushel-mito-nuc-SNP_22sep2020.RData
#> dietrichmemfeb2021_23feb2021   inProgress_dietrichmemfeb2021_23feb2021.RData
#> DMdata_07jan2021                           inProgress_DMdata_07jan2021.RData
#> DMdata_07oct2020                           inProgress_DMdata_07oct2020.RData
#> DMdata_11feb2021                           inProgress_DMdata_11feb2021.RData
#> DMdata_16dec2020                           inProgress_DMdata_16dec2020.RData
#> gokeyH1PD29may2019_27jan2021   inProgress_gokeyH1PD29may2019_27jan2021.RData
#> gokeyH1PD29may2019_28jan2021   inProgress_gokeyH1PD29may2019_28jan2021.RData
#> LottyDah_12jun2020                       inProgress_LottyDah_12jun2020.RData
#> proteinatlas_24jul2020               inProgress_proteinatlas_24jul2020.RData
#> ritux-workflow_10aug2020           inProgress_ritux-workflow_10aug2020.RData
#> SARSinteractome_24mar2020         inProgress_SARSinteractome_24mar2020.RData
#> SARSinteractome_26mar2020         inProgress_SARSinteractome_26mar2020.RData
#> SARSinteractome_27mar2020         inProgress_SARSinteractome_27mar2020.RData
#> slicejam-testing_07jul2020       inProgress_slicejam-testing_07jul2020.RData
#> testOnly_03jun2020                       inProgress_testOnly_03jun2020.RData
#> 
#> $session
#>  [1] "adam19-rnaseq"       "bushel-mito-nuc-SNP" "dietrichmemfeb2021" 
#>  [4] "DMdata"              "gokeyH1PD29may2019"  "LottyDah"           
#>  [7] "proteinatlas"        "ritux-workflow"      "SARSinteractome"    
#> [10] "slicejam-testing"    "testOnly"           
#>