List saved R sessions

List saved R sessions

list_jamsessions(
  sessions_path = jamsession_paths()$sessions,
  most_recent = FALSE,
  session_prefix = "^inProgress_",
  session_suffix = "[.]RData$",
  add_stats = FALSE,
  ...
)

Arguments

sessions_path
character vector of one or more file paths to search for saved R sessions. When NULL, it uses the output from jamsession_paths()$sessions.

most_recent
logical whether to return only the most recent saved entry for each session name. When most_recent is FALSE, this function returns all saved versions for all R session names.

session_prefix, session_suffix
character string used as prefix or suffix when searching each path in sessions_path for matching file names.

add_stats
logical indicating whether to include summary stats for each session: number_saved is the number of past versions of the session; total_size is the total file size for all saved versions.

...
additional parameters are ignored

Value

list containing

session_df

data.frame of R session information

session

vector of matching R session names

Details

This function lists available saved R sessions, usually called by grep_jamsessions().

When most_recent=TRUE it returns only the most recent session based upon the save_date.

It is possible to supply multiple file paths using the argument sessions_path, in which case all sessions are returned for all directories, then they are sorted by save_date. The input sessions_path is stored as an ordered factor, such that two sessions with the same save_date will still be ordered consistent with the input directories in sessions_path. In this case, when most_recent=TRUE there will be one entry returned per object name, and the first entry will be the most recent entry ordered by save_date then sessions_path. In practice, it is recommended to use one sessions_path location, but this option exists.

For example, jamsession(grep_sessions("my_session")) will load the most recent session by save_date where any ties prefer the order in sessions_path. Similarly, load_jamsession("my_session") does the same thing, without the grep_patten search step.

See also

Examples

list_jamsessions();

#> $session_df
#>                                   session save_date days_old file_size
#> LottyDah_12jun2020               LottyDah 12jun2020        0    1.1 kb
#> SARSinteractome_24mar2020 SARSinteractome 24mar2020       80      7 Mb
#> SARSinteractome_26mar2020 SARSinteractome 26mar2020       78     12 Mb
#> SARSinteractome_27mar2020 SARSinteractome 27mar2020       77     12 Mb
#> testOnly_03jun2020               testOnly 03jun2020        9     12 Mb
#>                                                session_path
#> LottyDah_12jun2020        /Users/wardjm/Projects/R-sessions
#> SARSinteractome_24mar2020 /Users/wardjm/Projects/R-sessions
#> SARSinteractome_26mar2020 /Users/wardjm/Projects/R-sessions
#> SARSinteractome_27mar2020 /Users/wardjm/Projects/R-sessions
#> testOnly_03jun2020        /Users/wardjm/Projects/R-sessions
#>                                                         session_file
#> LottyDah_12jun2020               inProgress_LottyDah_12jun2020.RData
#> SARSinteractome_24mar2020 inProgress_SARSinteractome_24mar2020.RData
#> SARSinteractome_26mar2020 inProgress_SARSinteractome_26mar2020.RData
#> SARSinteractome_27mar2020 inProgress_SARSinteractome_27mar2020.RData
#> testOnly_03jun2020               inProgress_testOnly_03jun2020.RData
#> 
#> $session
#> [1] "LottyDah"        "SARSinteractome" "testOnly"       
#>