list saved R objects

list saved R objects

list_objects(
  objects_path = jamsession_paths()$objects,
  most_recent = FALSE,
  object_prefix = "",
  object_suffix = "[.]RData$",
  add_stats = FALSE,
  ...
)

Arguments

objects_path
character vector of one or more file paths to search for saved R objects. When objects_path=NULL, it uses the output from jamsession_paths()$objects.

most_recent
logical whether to return only the most recent saved entry for each session name. When most_recent is FALSE, this function returns all saved versions for all R session names.

object_prefix, object_suffix
character string used as prefix or suffix when searching each path in objects_path for matching file names. By default, files are expected to have file extension .RData.

add_stats
logical indicating whether to include summary stats for each object.

...
additional arguments are ignored.

Value

list containing

object_df

data.frame of R objects

object

vector of matching R object names

Details

This function lists saved R objects for the given objects_path, usually called by grep_objects().

When most_recent=TRUE it returns only the most recent version of each file.

It is possible to supply multiple file paths using the argument objects_path, in which case all objects are returned for all directories, then they are sorted by save_date. The input objects_path is stored as an ordered factor, such that two files with the same save_date will still be ordered consistent with the input directories in objects_path. In this case, when most_recent=TRUE there will be one entry returned per object name, and the first entry will be the most recent entry ordered by save_date then objects_path. In practice, it is recommended to use one objects_path location, but this option exists.

For example, load_object(grep_objects("my_favorite_object")) will load the most recent object by save_date where any ties prefer the order in objects_path. Similarly, load_object("my_favorite_object") does the same thing, without the grep_patten search step.

See also

Examples

# by default the full list includes all past versions
head(list_objects());

#> $object_df
#>                                                          object save_date
#> arris_df_01jun2020                                     arris_df 01jun2020
#> arris_df_02jun2020                                     arris_df 02jun2020
#> arris_df_03jun2020                                     arris_df 03jun2020
#> arris_df_04jun2020                                     arris_df 04jun2020
#> arris_df_05jun2020                                     arris_df 05jun2020
#> arris_df_05may2020                                     arris_df 05may2020
#> arris_df_06jun2020                                     arris_df 06jun2020
#> arris_df_06may2020                                     arris_df 06may2020
#> arris_df_07jun2020                                     arris_df 07jun2020
#> arris_df_07may2020                                     arris_df 07may2020
#> arris_df_08jun2020                                     arris_df 08jun2020
#> arris_df_08may2020                                     arris_df 08may2020
#> arris_df_09jun2020                                     arris_df 09jun2020
#> arris_df_09may2020                                     arris_df 09may2020
#> arris_df_10jun2020                                     arris_df 10jun2020
#> arris_df_10may2020                                     arris_df 10may2020
#> arris_df_11jun2020                                     arris_df 11jun2020
#> arris_df_11may2020                                     arris_df 11may2020
#> arris_df_12may2020                                     arris_df 12may2020
#> arris_df_13may2020                                     arris_df 13may2020
#> arris_df_14may2020                                     arris_df 14may2020
#> arris_df_15may2020                                     arris_df 15may2020
#> arris_df_16may2020                                     arris_df 16may2020
#> arris_df_17may2020                                     arris_df 17may2020
#> arris_df_18may2020                                     arris_df 18may2020
#> arris_df_19may2020                                     arris_df 19may2020
#> arris_df_20may2020                                     arris_df 20may2020
#> arris_df_21may2020                                     arris_df 21may2020
#> arris_df_22may2020                                     arris_df 22may2020
#> arris_df_23may2020                                     arris_df 23may2020
#> arris_df_24may2020                                     arris_df 24may2020
#> arris_df_25may2020                                     arris_df 25may2020
#> arris_df_26may2020                                     arris_df 26may2020
#> arris_df_27may2020                                     arris_df 27may2020
#> arris_df_28may2020                                     arris_df 28may2020
#> arris_df_29may2020                                     arris_df 29may2020
#> arris_df_30may2020                                     arris_df 30may2020
#> arris_df_31may2020                                     arris_df 31may2020
#> arris_df-irow_11jun2020                           arris_df-irow 11jun2020
#> dietrich_track_lines_l_10apr2020         dietrich_track_lines_l 10apr2020
#> dietrichTxiNames_20mar2020                     dietrichTxiNames 20mar2020
#> dietrichTxis_20mar2020                             dietrichTxis 20mar2020
#> mem_dmlupus_affysoma_plots_12apr2020 mem_dmlupus_affysoma_plots 12apr2020
#> my_df_12jun2020                                           my_df 12jun2020
#> my_df-my_df2_12jun2020                             my_df-my_df2 12jun2020
#>                                      days_old file_size
#> arris_df_01jun2020                         11    184 kb
#> arris_df_02jun2020                         10    190 kb
#> arris_df_03jun2020                          9    200 kb
#> arris_df_04jun2020                          8    206 kb
#> arris_df_05jun2020                          7    213 kb
#> arris_df_05may2020                         38    6.3 kb
#> arris_df_06jun2020                          6    236 kb
#> arris_df_06may2020                         37     18 kb
#> arris_df_07jun2020                          5    257 kb
#> arris_df_07may2020                         36     24 kb
#> arris_df_08jun2020                          4    276 kb
#> arris_df_08may2020                         35     31 kb
#> arris_df_09jun2020                          3    282 kb
#> arris_df_09may2020                         34     41 kb
#> arris_df_10jun2020                          2    294 kb
#> arris_df_10may2020                         33     48 kb
#> arris_df_11jun2020                          1    298 kb
#> arris_df_11may2020                         32     60 kb
#> arris_df_12may2020                         31     64 kb
#> arris_df_13may2020                         30     69 kb
#> arris_df_14may2020                         29     76 kb
#> arris_df_15may2020                         28     83 kb
#> arris_df_16may2020                         27     92 kb
#> arris_df_17may2020                         26    100 kb
#> arris_df_18may2020                         25    103 kb
#> arris_df_19may2020                         24    109 kb
#> arris_df_20may2020                         23    113 kb
#> arris_df_21may2020                         22    117 kb
#> arris_df_22may2020                         21    122 kb
#> arris_df_23may2020                         20    127 kb
#> arris_df_24may2020                         19    135 kb
#> arris_df_25may2020                         18    143 kb
#> arris_df_26may2020                         17    151 kb
#> arris_df_27may2020                         16    159 kb
#> arris_df_28may2020                         15    161 kb
#> arris_df_29may2020                         14    162 kb
#> arris_df_30may2020                         13    170 kb
#> arris_df_31may2020                         12    183 kb
#> arris_df-irow_11jun2020                     1    294 kb
#> dietrich_track_lines_l_10apr2020           63      1 kb
#> dietrichTxiNames_20mar2020                 84 345 bytes
#> dietrichTxis_20mar2020                     84     51 Mb
#> mem_dmlupus_affysoma_plots_12apr2020       61    299 kb
#> my_df_12jun2020                             0 160 bytes
#> my_df-my_df2_12jun2020                      0 212 bytes
#>                                                           object_path
#> arris_df_01jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_02jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_03jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_04jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_05jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_05may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_06jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_06may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_07jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_07may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_08jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_08may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_09jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_09may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_10jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_10may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_11jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df_11may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_12may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_13may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_14may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_15may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_16may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_17may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_18may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_19may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_20may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_21may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_22may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_23may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_24may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_25may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_26may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_27may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_28may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_29may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_30may2020                   /Users/wardjm/Projects/R-objects
#> arris_df_31may2020                   /Users/wardjm/Projects/R-objects
#> arris_df-irow_11jun2020              /Users/wardjm/Projects/R-objects
#> dietrich_track_lines_l_10apr2020     /Users/wardjm/Projects/R-objects
#> dietrichTxiNames_20mar2020           /Users/wardjm/Projects/R-objects
#> dietrichTxis_20mar2020               /Users/wardjm/Projects/R-objects
#> mem_dmlupus_affysoma_plots_12apr2020 /Users/wardjm/Projects/R-objects
#> my_df_12jun2020                      /Users/wardjm/Projects/R-objects
#> my_df-my_df2_12jun2020               /Users/wardjm/Projects/R-objects
#>                                                                     object_file
#> arris_df_01jun2020                                     arris_df_01jun2020.RData
#> arris_df_02jun2020                                     arris_df_02jun2020.RData
#> arris_df_03jun2020                                     arris_df_03jun2020.RData
#> arris_df_04jun2020                                     arris_df_04jun2020.RData
#> arris_df_05jun2020                                     arris_df_05jun2020.RData
#> arris_df_05may2020                                     arris_df_05may2020.RData
#> arris_df_06jun2020                                     arris_df_06jun2020.RData
#> arris_df_06may2020                                     arris_df_06may2020.RData
#> arris_df_07jun2020                                     arris_df_07jun2020.RData
#> arris_df_07may2020                                     arris_df_07may2020.RData
#> arris_df_08jun2020                                     arris_df_08jun2020.RData
#> arris_df_08may2020                                     arris_df_08may2020.RData
#> arris_df_09jun2020                                     arris_df_09jun2020.RData
#> arris_df_09may2020                                     arris_df_09may2020.RData
#> arris_df_10jun2020                                     arris_df_10jun2020.RData
#> arris_df_10may2020                                     arris_df_10may2020.RData
#> arris_df_11jun2020                                     arris_df_11jun2020.RData
#> arris_df_11may2020                                     arris_df_11may2020.RData
#> arris_df_12may2020                                     arris_df_12may2020.RData
#> arris_df_13may2020                                     arris_df_13may2020.RData
#> arris_df_14may2020                                     arris_df_14may2020.RData
#> arris_df_15may2020                                     arris_df_15may2020.RData
#> arris_df_16may2020                                     arris_df_16may2020.RData
#> arris_df_17may2020                                     arris_df_17may2020.RData
#> arris_df_18may2020                                     arris_df_18may2020.RData
#> arris_df_19may2020                                     arris_df_19may2020.RData
#> arris_df_20may2020                                     arris_df_20may2020.RData
#> arris_df_21may2020                                     arris_df_21may2020.RData
#> arris_df_22may2020                                     arris_df_22may2020.RData
#> arris_df_23may2020                                     arris_df_23may2020.RData
#> arris_df_24may2020                                     arris_df_24may2020.RData
#> arris_df_25may2020                                     arris_df_25may2020.RData
#> arris_df_26may2020                                     arris_df_26may2020.RData
#> arris_df_27may2020                                     arris_df_27may2020.RData
#> arris_df_28may2020                                     arris_df_28may2020.RData
#> arris_df_29may2020                                     arris_df_29may2020.RData
#> arris_df_30may2020                                     arris_df_30may2020.RData
#> arris_df_31may2020                                     arris_df_31may2020.RData
#> arris_df-irow_11jun2020                           arris_df-irow_11jun2020.RData
#> dietrich_track_lines_l_10apr2020         dietrich_track_lines_l_10apr2020.RData
#> dietrichTxiNames_20mar2020                     dietrichTxiNames_20mar2020.RData
#> dietrichTxis_20mar2020                             dietrichTxis_20mar2020.RData
#> mem_dmlupus_affysoma_plots_12apr2020 mem_dmlupus_affysoma_plots_12apr2020.RData
#> my_df_12jun2020                                           my_df_12jun2020.RData
#> my_df-my_df2_12jun2020                             my_df-my_df2_12jun2020.RData
#> 
#> $object
#> NULL
#> 

# you can return only the most recent for each object
list_objects(most_recent=TRUE);

#> $object_df
#>                                                          object save_date
#> arris_df_11jun2020                                     arris_df 11jun2020
#> arris_df-irow_11jun2020                           arris_df-irow 11jun2020
#> dietrich_track_lines_l_10apr2020         dietrich_track_lines_l 10apr2020
#> dietrichTxiNames_20mar2020                     dietrichTxiNames 20mar2020
#> dietrichTxis_20mar2020                             dietrichTxis 20mar2020
#> mem_dmlupus_affysoma_plots_12apr2020 mem_dmlupus_affysoma_plots 12apr2020
#> my_df_12jun2020                                           my_df 12jun2020
#> my_df-my_df2_12jun2020                             my_df-my_df2 12jun2020
#>                                      days_old file_size
#> arris_df_11jun2020                          1    298 kb
#> arris_df-irow_11jun2020                     1    294 kb
#> dietrich_track_lines_l_10apr2020           63      1 kb
#> dietrichTxiNames_20mar2020                 84 345 bytes
#> dietrichTxis_20mar2020                     84     51 Mb
#> mem_dmlupus_affysoma_plots_12apr2020       61    299 kb
#> my_df_12jun2020                             0 160 bytes
#> my_df-my_df2_12jun2020                      0 212 bytes
#>                                                           object_path
#> arris_df_11jun2020                   /Users/wardjm/Projects/R-objects
#> arris_df-irow_11jun2020              /Users/wardjm/Projects/R-objects
#> dietrich_track_lines_l_10apr2020     /Users/wardjm/Projects/R-objects
#> dietrichTxiNames_20mar2020           /Users/wardjm/Projects/R-objects
#> dietrichTxis_20mar2020               /Users/wardjm/Projects/R-objects
#> mem_dmlupus_affysoma_plots_12apr2020 /Users/wardjm/Projects/R-objects
#> my_df_12jun2020                      /Users/wardjm/Projects/R-objects
#> my_df-my_df2_12jun2020               /Users/wardjm/Projects/R-objects
#>                                                                     object_file
#> arris_df_11jun2020                                     arris_df_11jun2020.RData
#> arris_df-irow_11jun2020                           arris_df-irow_11jun2020.RData
#> dietrich_track_lines_l_10apr2020         dietrich_track_lines_l_10apr2020.RData
#> dietrichTxiNames_20mar2020                     dietrichTxiNames_20mar2020.RData
#> dietrichTxis_20mar2020                             dietrichTxis_20mar2020.RData
#> mem_dmlupus_affysoma_plots_12apr2020 mem_dmlupus_affysoma_plots_12apr2020.RData
#> my_df_12jun2020                                           my_df_12jun2020.RData
#> my_df-my_df2_12jun2020                             my_df-my_df2_12jun2020.RData
#> 
#> $object
#> NULL
#> 

# add_stats makes sense when most_recent=TRUE
list_objects(most_recent=TRUE, add_stats=TRUE);

#> $object_df
#>                       object save_date days_old file_size
#> 1                   arris_df 11jun2020        1    298 kb
#> 2              arris_df-irow 11jun2020        1    294 kb
#> 3     dietrich_track_lines_l 10apr2020       63      1 kb
#> 4           dietrichTxiNames 20mar2020       84 345 bytes
#> 5               dietrichTxis 20mar2020       84     51 Mb
#> 6 mem_dmlupus_affysoma_plots 12apr2020       61    299 kb
#> 7                      my_df 12jun2020        0 160 bytes
#> 8               my_df-my_df2 12jun2020        0 212 bytes
#>                        object_path                                object_file
#> 1 /Users/wardjm/Projects/R-objects                   arris_df_11jun2020.RData
#> 2 /Users/wardjm/Projects/R-objects              arris_df-irow_11jun2020.RData
#> 3 /Users/wardjm/Projects/R-objects     dietrich_track_lines_l_10apr2020.RData
#> 4 /Users/wardjm/Projects/R-objects           dietrichTxiNames_20mar2020.RData
#> 5 /Users/wardjm/Projects/R-objects               dietrichTxis_20mar2020.RData
#> 6 /Users/wardjm/Projects/R-objects mem_dmlupus_affysoma_plots_12apr2020.RData
#> 7 /Users/wardjm/Projects/R-objects                      my_df_12jun2020.RData
#> 8 /Users/wardjm/Projects/R-objects               my_df-my_df2_12jun2020.RData
#>   number_saved total_size days_span
#> 1           38     5.2 Mb        38
#> 2            1     294 kb         1
#> 3            1       1 kb         1
#> 4            1  345 bytes         1
#> 5            1      51 Mb         1
#> 6            1     299 kb         1
#> 7            1  160 bytes         1
#> 8            1  212 bytes         1
#> 
#> $object
#> NULL
#>